refactor: 오버엔지니어링으로 인해 추상화를 구현체로 변경하라
