--- src/test/java/stringcalculator/factory/splitter/Splitter.java
@@ -1,9 +1,0 @@
-package stringcalculator.factory.splitter;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface Splitter {
-
-    List<String> split(final String value);
-}
